+++ docs/assets/modals.stories-BXwjsgWV.js
@@ -1,0 +1,10 @@
+import{_ as t}from"./UluButton-B0ZbLSM-.js";import"./iframe-BSWppwhy.js";import"./preload-helper-BJwshlQW.js";import"./UluAction-DFdM0w1k.js";import"./props-DEaRQ31f.js";const r={title:"Plugins/modals"},e=()=>({components:{UluButton:t},template:`
+    <UluButton @click="$uluModals.open('demo')" text="Open App-wide Modal"/>
+  `}),o=e.bind({});o.args={};o.parameters={...o.parameters,docs:{...o.parameters?.docs,source:{originalSource:`() => ({
+  components: {
+    UluButton
+  },
+  "template": \`
+    <UluButton @click="$uluModals.open('demo')" text="Open App-wide Modal"/>
+  \`
+})`,...o.parameters?.docs?.source}}};const m=["GlobalApiTest"];export{o as GlobalApiTest,m as __namedExportsOrder,r as default};