+++ lib/components/elements/UluButton.vue
@@ -72,7 +72,8 @@
      */
     download: [Boolean, String],
     /**
-     * @deprecated Use `ariaLabel` instead. For icon only buttons or buttons that need an explicit label
+     * **Deprecated!** Use `ariaLabel` instead. 
+     * @deprecated Use `ariaLabel` instead.
      */
     alt: String,
     /**