+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.6.1
+
+- **Forms** remove deprecated prop warnings and remove deprecated props (ie. label, required, etc that were on the fields and are now on UluFormItem)
+  - Keeping them is more confusing in storybook docs
+
 ## 0.6.0
 
 - **Forms Updates (Breaking Changes)**